move verify_message_signatures
