+++ anchor/message_validator/src/consensus_message.rs
@@ -1,18 +1,15 @@
 use crate::consensus_state::ConsensusState;
 use crate::{
-    compute_quorum_size, hash_data, ValidatedSSVMessage, ValidationContext, ValidationFailure,
+    compute_quorum_size, hash_data, verify_message_signatures, ValidatedSSVMessage,
+    ValidationContext, ValidationFailure,
 };
-use openssl::hash::MessageDigest;
-use openssl::pkey::{PKey, Public};
-use openssl::rsa::Rsa;
-use openssl::sign::Verifier;
 use slot_clock::SlotClock;
 use ssv_types::consensus::{QbftMessage, QbftMessageType};
 use ssv_types::message::SignedSSVMessage;
 use ssv_types::msgid::Role;
 use ssv_types::{CommitteeInfo, IndexSet, OperatorId, VariableList};
 use ssv_types::{Round, Slot};
-use ssz::{Decode, Encode};
+use ssz::Decode;
 use std::convert::Into;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
@@ -256,57 +253,6 @@ pub(crate) fn validate_qbft_logic(
     Ok(())
 }
 
-/// Verifies all signatures in a signed SSV message
-fn verify_message_signatures(
-    signed_message: &SignedSSVMessage,
-    operators_pks: &[Rsa<Public>],
-) -> Result<(), ValidationFailure> {
-    let signatures = signed_message.signatures();
-
-    // Basic validation for signature/operator count matching
-    if signatures.len() != operators_pks.len() {
-        return Err(ValidationFailure::SignatureVerificationFailed {
-            reason: "Signature count doesn't match operator count".to_string(),
-        });
-    }
-
-    for (signature, operators_pk) in signatures.iter().zip(operators_pks.iter()) {
-        let p_key = PKey::from_rsa(operators_pk.clone()).map_err(|e| {
-            ValidationFailure::SignatureVerificationFailed {
-                reason: format!("Failed to create PKey: {}", e),
-            }
-        })?;
-
-        let mut verifier = Verifier::new(MessageDigest::sha256(), &p_key).map_err(|e| {
-            ValidationFailure::SignatureVerificationFailed {
-                reason: format!("Failed to create verifier: {}", e),
-            }
-        })?;
-
-        verifier
-            .update(&signed_message.ssv_message().as_ssz_bytes())
-            .map_err(|e| ValidationFailure::SignatureVerificationFailed {
-                reason: format!("Failed to update verifier: {}", e),
-            })?;
-
-        match verifier.verify(signature) {
-            Ok(true) => {}
-            Ok(false) => {
-                return Err(ValidationFailure::SignatureVerificationFailed {
-                    reason: "Signature verification failed".to_string(),
-                });
-            }
-            Err(e) => {
-                return Err(ValidationFailure::SignatureVerificationFailed {
-                    reason: format!("Signature verification error: {}", e),
-                });
-            }
-        }
-    }
-
-    Ok(())
-}
-
 // Define constants to match the Go implementation
 const FIRST_ROUND: u64 = 1;
 const MAX_ALLOWED_ROUNDS_FUTURE: u64 = 3;
@@ -412,6 +358,7 @@ mod tests {
     };
     use crate::{validate_ssv_message, ValidatedSSVMessage};
     use bls::{Hash256, PublicKeyBytes};
+    use openssl::hash::MessageDigest;
     use slot_clock::ManualSlotClock;
     use ssv_types::consensus::{QbftMessage, QbftMessageType};
     use ssv_types::domain_type::DomainType;
